Rename `CompileError.h` to `Diagnostics.h`
