+++ Source/SpireCore/Diagnostics.h
