+++ Source/SpireCore/ShaderCompiler.h
@@ -2,7 +2,7 @@
 #define RASTER_SHADER_COMPILER_H
 
 #include "../CoreLib/Basic.h"
-#include "CompileError.h"
+#include "Diagnostics.h"
 #include "CompiledProgram.h"
 #include "Syntax.h"
 #include "CodeGenBackend.h"
