+++ Source/SpireCore/SpireCore.vcxproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup Label="ProjectConfigurations">
     <ProjectConfiguration Include="DebugClang|Win32">
@@ -373,7 +373,7 @@
     <ClInclude Include="CodeGenBackend.h" />
     <ClInclude Include="CodeWriter.h" />
     <ClInclude Include="CompiledProgram.h" />
-    <ClInclude Include="CompileError.h" />
+    <ClInclude Include="Diagnostics.h" />
     <ClInclude Include="GetDependencyVisitor.h" />
     <ClInclude Include="Naming.h" />
     <ClInclude Include="Schedule.h" />
