+++ Source/SpireCore/SpireCore.vcxproj.filters
@@ -9,9 +9,6 @@
     </Filter>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="CompileError.h">
-      <Filter>Front End</Filter>
-    </ClInclude>
     <ClInclude Include="Lexer.h">
       <Filter>Front End</Filter>
     </ClInclude>
@@ -70,6 +67,9 @@
       <Filter>Back End</Filter>
     </ClInclude>
     <ClInclude Include="GetDependencyVisitor.h">
+      <Filter>Front End</Filter>
+    </ClInclude>
+    <ClInclude Include="Diagnostics.h">
       <Filter>Front End</Filter>
     </ClInclude>
   </ItemGroup>
